refactor(API types): reorganize types
