+++ src/api/get/DeviceGet.ts
