+++ src/api/get/DevicePowerGet.ts
